refactor: Move InstructorFactory helper to common/
