+++ cms/djangoapps/contentstore/rest_api/v1/tests/test_views.py
@@ -9,8 +9,8 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 
 from common.djangoapps.student.tests.factories import GlobalStaffFactory
+from common.djangoapps.student.tests.factories import InstructorFactory
 from common.djangoapps.student.tests.factories import UserFactory
-from lms.djangoapps.courseware.tests.factories import InstructorFactory
 from xmodule.modulestore.django import modulestore
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory
