+++ common/djangoapps/student/tests/test_roles.py
@@ -8,7 +8,6 @@ import six
 from django.test import TestCase
 from opaque_keys.edx.keys import CourseKey
 
-from lms.djangoapps.courseware.tests.factories import InstructorFactory
 from common.djangoapps.student.roles import (
     CourseBetaTesterRole,
     CourseInstructorRole,
@@ -20,6 +19,7 @@ from common.djangoapps.student.roles import (
     RoleCache
 )
 from common.djangoapps.student.tests.factories import AnonymousUserFactory
+from common.djangoapps.student.tests.factories import InstructorFactory
 from common.djangoapps.student.tests.factories import StaffFactory
 from common.djangoapps.student.tests.factories import UserFactory
