+++ lms/djangoapps/course_wiki/tests/test_comprehensive_theming.py
@@ -8,8 +8,8 @@ from unittest import skip
 from django.test.client import Client
 from wiki.models import URLPath
 
+from common.djangoapps.student.tests.factories import InstructorFactory
 from lms.djangoapps.course_wiki.views import get_or_create_root
-from lms.djangoapps.courseware.tests.factories import InstructorFactory
 from openedx.core.djangoapps.theming.tests.test_util import with_comprehensive_theme
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory
