+++ lms/djangoapps/instructor/tests/test_api.py
@@ -54,6 +54,7 @@ from common.djangoapps.student.roles import (
 )
 from common.djangoapps.student.tests.factories import CourseEnrollmentFactory
 from common.djangoapps.student.tests.factories import GlobalStaffFactory
+from common.djangoapps.student.tests.factories import InstructorFactory
 from common.djangoapps.student.tests.factories import StaffFactory
 from common.djangoapps.student.tests.factories import UserFactory
 from lms.djangoapps.bulk_email.models import BulkEmailFlag, CourseEmail, CourseEmailTemplate
@@ -65,7 +66,6 @@ from lms.djangoapps.certificates.tests.factories import (
 from lms.djangoapps.courseware.models import StudentModule
 from lms.djangoapps.courseware.tests.factories import (
     BetaTesterFactory,
-    InstructorFactory,
 )
 from lms.djangoapps.courseware.tests.helpers import LoginEnrollmentTestCase
 from lms.djangoapps.instructor.tests.utils import FakeContentTask, FakeEmail, FakeEmailInfo
