+++ openedx/features/course_duration_limits/tests/test_course_expiration.py
@@ -15,10 +15,10 @@ from common.djangoapps.student.models import CourseEnrollment, FBEEnrollmentExcl
 from common.djangoapps.student.roles import CourseInstructorRole
 from common.djangoapps.student.tests.factories import TEST_PASSWORD, CourseEnrollmentFactory, UserFactory
 from common.djangoapps.student.tests.factories import GlobalStaffFactory
+from common.djangoapps.student.tests.factories import InstructorFactory
 from common.djangoapps.student.tests.factories import StaffFactory
 from lms.djangoapps.courseware.tests.factories import (
     BetaTesterFactory,
-    InstructorFactory,
     OrgInstructorFactory,
     OrgStaffFactory,
 )
